+++ GRDB/Core/DatabasePool.swift
@@ -366,54 +366,6 @@ extension DatabasePool: DatabaseReader {
             }
     }
     
-    /// :nodoc:
-    public func _weakAsyncRead(_ value: @escaping (Result<Database, Error>?) -> Void) {
-        // First async jump in order to grab a reader connection.
-        // Honor configuration dispatching (qos/targetQueue).
-        let label = configuration.identifier(
-            defaultLabel: "GRDB.DatabasePool",
-            purpose: "asyncRead")
-        configuration
-            .makeReaderDispatchQueue(label: label)
-            .async { [weak self] in
-                guard let self = self else {
-                    value(nil)
-                    return
-                }
-                
-                do {
-                    guard let readerPool = self.readerPool else {
-                        throw DatabaseError.connectionIsClosed()
-                    }
-                    let (reader, releaseReader) = try readerPool.get()
-                    
-                    // Second async jump because sync could deadlock if
-                    // configuration has a serial targetQueue.
-                    reader.weakAsync { db in
-                        guard let db = db else {
-                            value(nil)
-                            return
-                        }
-                        
-                        defer {
-                            try? db.commit() // Ignore commit error
-                            releaseReader()
-                        }
-                        do {
-                            // The block isolation comes from the DEFERRED transaction.
-                            try db.beginTransaction(.deferred)
-                            try db.clearSchemaCacheIfNeeded()
-                            value(.success(db))
-                        } catch {
-                            value(.failure(error))
-                        }
-                    }
-                } catch {
-                    value(.failure(error))
-                }
-            }
-    }
-    
     @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func unsafeRead<T>(_ value: (Database) throws -> T) throws -> T {
         GRDBPrecondition(currentReader == nil, "Database methods are not reentrant.")
@@ -748,11 +700,6 @@ extension DatabasePool: DatabaseReader {
         writer.async(updates)
     }
     
-    /// :nodoc:
-    public func _weakAsyncWriteWithoutTransaction(_ updates: @escaping (Database?) -> Void) {
-        writer.weakAsync(updates)
-    }
-    
     // MARK: - Database Observation
     
     /// :nodoc:
@@ -763,25 +710,26 @@ extension DatabasePool: DatabaseReader {
     -> DatabaseCancellable
     {
         if configuration.readonly {
+            // The easy case: the database does not change
             return _addReadOnly(
                 observation: observation,
                 scheduling: scheduler,
                 onChange: onChange)
-        }
-        
-        if observation.requiresWriteAccess {
-            let observer = _addWriteOnly(
+            
+        } else if observation.requiresWriteAccess {
+            // Observe from the writer database connection.
+            return _addWriteOnly(
                 observation: observation,
                 scheduling: scheduler,
                 onChange: onChange)
-            return AnyDatabaseCancellable(cancel: observer.cancel)
+            
+        } else {
+            // DatabasePool can perform concurrent observation
+            return _addConcurrent(
+                observation: observation,
+                scheduling: scheduler,
+                onChange: onChange)
         }
-        
-        let observer = _addConcurrent(
-            observation: observation,
-            scheduling: scheduler,
-            onChange: onChange)
-        return AnyDatabaseCancellable(cancel: observer.cancel)
     }
     
     /// A concurrent observation fetches the initial value without waiting for
@@ -790,178 +738,19 @@ extension DatabasePool: DatabaseReader {
         observation: ValueObservation<Reducer>,
         scheduling scheduler: ValueObservationScheduler,
         onChange: @escaping (Reducer.Value) -> Void)
-    -> ValueObserver<Reducer> // For testability
+    -> DatabaseCancellable
     {
         assert(!configuration.readonly, "Use _addReadOnly(observation:) instead")
         assert(!observation.requiresWriteAccess, "Use _addWriteOnly(observation:) instead")
-        
-        let reduceQueueLabel = configuration.identifier(
-            defaultLabel: "GRDB",
-            purpose: "ValueObservation")
-        let observer = ValueObserver(
-            observation: observation,
-            writer: self,
+        let observer = ValueConcurrentObserver(
+            dbPool: self,
             scheduler: scheduler,
-            reduceQueue: configuration.makeReaderDispatchQueue(label: reduceQueueLabel),
+            trackingMode: observation.trackingMode,
+            reducer: observation.makeReducer(),
+            events: observation.events,
             onChange: onChange)
-        
-        // Starting a concurrent observation means that we'll fetch the initial
-        // value right away, without waiting for an access to the writer queue,
-        // and the opportunity to install a transaction observer.
-        //
-        // This is how DatabasePool can start an observation and promptly notify
-        // an initial value, even when there is a long-running write transaction
-        // in the background.
-        //
-        // We thus have to deal with the fact that between this initial fetch,
-        // and the beginning of transaction tracking, any number of untracked
-        // writes may occur.
-        //
-        // We must notify changes that happen during this untracked window. But
-        // how do we spot them, since we're not tracking database changes yet?
-        //
-        // A safe solution is to always perform a second fetch. We may fetch the
-        // same initial value, if no change did happen. But we don't miss
-        // possible changes.
-        //
-        // We can avoid this second fetch when SQLite is compiled with the
-        // SQLITE_ENABLE_SNAPSHOT option:
-        //
-        // 1. Perform the initial fetch in a DatabaseSnapshot. Its long running
-        // transaction acquires a lock that will prevent checkpointing until we
-        // get a writer access, so that we can reliably compare database
-        // versions with `sqlite3_snapshot`:
-        // https://www.sqlite.org/c3ref/snapshot.html.
-        //
-        // 2. Get a writer access, and compare the versions of the initial
-        // snapshot, and the current state of the database: if versions are
-        // identical, we can avoid the second fetch. If they are not, we perform
-        // the second fetch, even if the actual changes are unrelated to the
-        // tracked database region (we have no way to know).
-        //
-        // 3. Install the transaction observer.
-        
-        #if SQLITE_ENABLE_SNAPSHOT
-        if scheduler.immediateInitialValue() {
-            do {
-                let initialSnapshot = try makeSnapshot()
-                let initialValue = try initialSnapshot.read(observer.fetchInitialValue)
-                onChange(initialValue)
-                add(observer: observer, from: initialSnapshot)
-            } catch {
-                observer.complete()
-                observation.events.didFail?(error)
-            }
-        } else {
-            let label = configuration.identifier(
-                defaultLabel: "GRDB.DatabasePool",
-                purpose: "ValueObservation")
-            configuration
-                .makeReaderDispatchQueue(label: label)
-                .async { [weak self] in
-                    guard let self = self else { return }
-                    if observer.isCompleted { return }
-                    
-                    do {
-                        let initialSnapshot = try self.makeSnapshot()
-                        let initialValue = try initialSnapshot.read(observer.fetchInitialValue)
-                        observer.notifyChange(initialValue)
-                        self.add(observer: observer, from: initialSnapshot)
-                    } catch {
-                        observer.notifyErrorAndComplete(error)
-                    }
-                }
-        }
-        #else
-        if scheduler.immediateInitialValue() {
-            do {
-                let initialValue = try read(observer.fetchInitialValue)
-                onChange(initialValue)
-                addObserver(observer: observer)
-            } catch {
-                observer.complete()
-                observation.events.didFail?(error)
-            }
-        } else {
-            _weakAsyncRead { [weak self] dbResult in
-                guard let self = self, let dbResult = dbResult else { return }
-                if observer.isCompleted { return }
-                
-                do {
-                    let initialValue = try observer.fetchInitialValue(dbResult.get())
-                    observer.notifyChange(initialValue)
-                    self.addObserver(observer: observer)
-                } catch {
-                    observer.notifyErrorAndComplete(error)
-                }
-            }
-        }
-        #endif
-        
-        return observer
+        return observer.start()
     }
-    
-    #if SQLITE_ENABLE_SNAPSHOT
-    // Support for _addConcurrent(observation:)
-    private func add<Reducer: ValueReducer>(
-        observer: ValueObserver<Reducer>,
-        from initialSnapshot: DatabaseSnapshot)
-    {
-        _weakAsyncWriteWithoutTransaction { db in
-            guard let db = db else { return }
-            if observer.isCompleted { return }
-            
-            do {
-                // Transaction is needed for version snapshotting
-                try db.inTransaction(.deferred) {
-                    // Keep DatabaseSnaphot alive until we have compared
-                    // database versions. It prevents database checkpointing,
-                    // and keeps versions (`sqlite3_snapshot`) valid
-                    // and comparable.
-                    let fetchNeeded: Bool = try withExtendedLifetime(initialSnapshot) {
-                        guard let initialVersion = initialSnapshot.version else {
-                            return true
-                        }
-                        return try db.wasChanged(since: initialVersion)
-                    }
-                    
-                    if fetchNeeded {
-                        observer.events.databaseDidChange?()
-                        if let value = try observer.fetchValue(db) {
-                            observer.notifyChange(value)
-                        }
-                    }
-                    return .commit
-                }
-                
-                // Now we can start observation
-                db.add(transactionObserver: observer, extent: .observerLifetime)
-            } catch {
-                observer.notifyErrorAndComplete(error)
-            }
-        }
-    }
-    #else
-    // Support for _addConcurrent(observation:)
-    private func addObserver<Reducer: ValueReducer>(observer: ValueObserver<Reducer>) {
-        _weakAsyncWriteWithoutTransaction { db in
-            guard let db = db else { return }
-            if observer.isCompleted { return }
-            
-            do {
-                observer.events.databaseDidChange?()
-                if let value = try observer.fetchValue(db) {
-                    observer.notifyChange(value)
-                }
-                
-                // Now we can start observation
-                db.add(transactionObserver: observer, extent: .observerLifetime)
-            } catch {
-                observer.notifyErrorAndComplete(error)
-            }
-        }
-    }
-    #endif
 }
 
 extension DatabasePool {