+++ GRDB/Utils/Utils.swift
@@ -172,7 +172,7 @@ func concat<T>(_ rhs: ((T) -> Void)?, _ lhs: ((T) -> Void)?) -> ((T) -> Void)? {
     }
 }
 
-extension NSRecursiveLock {
+extension NSLocking {
     func synchronized<T>(
         _ message: @autoclosure () -> String = #function,
         _ block: () throws -> T)